non-hovering actions
